+++ eventstore/mongodb/spring/blocking/src/main/java/org/occurrent/eventstore/mongodb/spring/blocking/SpringMongoEventStore.java
@@ -335,10 +335,9 @@ public class SpringMongoEventStore implements EventStore, EventStoreOperations, 
             mongoTemplate.createCollection(eventStoreCollectionName);
         }
         MongoCollection<Document> eventStoreCollection = mongoTemplate.getCollection(eventStoreCollectionName);
-        eventStoreCollection.createIndex(Indexes.ascending(STREAM_ID));
         // Cloud spec defines id + source must be unique!
         eventStoreCollection.createIndex(Indexes.compoundIndex(Indexes.ascending("id"), Indexes.ascending("source")), new IndexOptions().unique(true));
-        // Create a streamId + streamVersion index
+        // Create a streamId + streamVersion index (note that we don't need to index stream id separately since it's covered by this compound index)
         eventStoreCollection.createIndex(Indexes.compoundIndex(Indexes.ascending(STREAM_ID), Indexes.descending(STREAM_VERSION)), new IndexOptions().unique(true));
 
         // SessionSynchronization need to be "ALWAYS" in order for TransactionTemplate to work with mongo template!